step1 : 이전 주차의 코드로 초기화
