+++ src/test/java/nextstep/acceptance/RestAssuredTest.java
@@ -32,47 +32,4 @@ public class RestAssuredTest {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 
-    static class AuthAcceptanceTest extends AcceptanceTest {
-        public static final String EMAIL = "admin@email.com";
-        public static final String PASSWORD = "password";
-        public static final Integer AGE = 20;
-
-        @Autowired
-        private MemberRepository memberRepository;
-
-        @DisplayName("Bearer Auth")
-        @Test
-        void bearerAuth() {
-            memberRepository.save(new Member(EMAIL, PASSWORD, AGE));
-
-            Map<String, String> params = new HashMap<>();
-            params.put("email", EMAIL);
-            params.put("password", PASSWORD);
-
-            ExtractableResponse<Response> response = RestAssured.given().log().all()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(params)
-                    .when().post("/login/token")
-                    .then().log().all()
-                    .statusCode(HttpStatus.OK.value()).extract();
-
-            assertThat(response.jsonPath().getString("accessToken")).isNotBlank();
-        }
-
-        @DisplayName("Github Auth")
-        @Test
-        void githubAuth() {
-            Map<String, String> params = new HashMap<>();
-            params.put("code", "code");
-
-            ExtractableResponse<Response> response = RestAssured.given().log().all()
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(params)
-                    .when().post("/login/github")
-                    .then().log().all()
-                    .statusCode(HttpStatus.OK.value()).extract();
-
-            assertThat(response.jsonPath().getString("accessToken")).isNotBlank();
-        }
-    }
 }
